Implement messenger_full problem.
Please note that the code still yields identical results to the one in the GTOPtoolbox; The calculated objective values are different than the values presented on the ESA problem page, but this difference was already in the toolbox itself and is not a result of adapting the code to PASS.

This commit only copies the contents of the GTOPToolbox with minimal adjustments. I already found and fixed one memory leak, but I wouldn't be too sure that there aren't any more. Further cleanup might be useful.